Sea-water properties

PhysOcean

Sea-water properties

PhysOcean.density — Method.

density(S,T,p)

Compute the density of sea-water (kg/m³) at the salinity S (psu, PSS-78), temperature T (degree Celsius, ITS-90) and pressure p (decibar) using the UNESCO 1983 polynomial.

Fofonoff, N.P.; Millard, R.C. (1983). Algorithms for computation of fundamental properties of seawater. UNESCO Technical Papers in Marine Science, No. 44. UNESCO: Paris. 53 pp. http://web.archive.org/web/20170103000527/http://unesdoc.unesco.org/images/0005/000598/059832eb.pdf

source

PhysOcean.secant_bulk_modulus — Method.

secant_bulk_modulus(S,T,p)

Compute the secant bulk modulus of sea-water (bars) at the salinity S (psu, PSS-78), temperature T (degree Celsius, ITS-90) and pressure p (decibar) using the UNESCO polynomial 1983.

Fofonoff, N.P.; Millard, R.C. (1983). Algorithms for computation of fundamental properties of seawater. UNESCO Technical Papers in Marine Science, No. 44. UNESCO: Paris. 53 pp. http://web.archive.org/web/20170103000527/http://unesdoc.unesco.org/images/0005/000598/059832eb.pdf

source

PhysOcean.freezing_temperature — Method.

freezing_temperature(S)

Compute the freezing temperature (in degree Celsius) of sea-water based on the salinity S (psu).

source

Heat fluxes

PhysOcean.latentflux — Method.

latentflux(Ts,Ta,r,w,pa)

Compute the latent heat flux (W/m²) using the sea surface temperature Ts (degree Celsius), the air temperature Ta (degree Celsius), the relative humidity r (0 ≤ r ≤ 1, pressure ratio, not percentage), the wind speed w (m/s) and the air pressure (hPa).

source

PhysOcean.longwaveflux — Method.

longwaveflux(Ts,Ta,e,tcc)

Compute the long wave heat flux (W/m²) using the sea surface temperature Ts (degree Celsius), the air temperature Ta (degree Celsius), the wate vapour pressure e (hPa) and the total cloud coverage ttc (0 ≤ tcc ≤ 1).

source

PhysOcean.sensibleflux — Method.

sensibleflux(Ts,Ta,w)

Compute the sensible heat flux (W/m²) using the wind speed w (m/s), the sea surface temperature Ts (degree Celsius), the air temperature Ta (degree Celsius).

source

PhysOcean.solarflux — Method.

solarflux(Q,al)

Compute the solar heat flux (W/m²)

source

PhysOcean.vaporpressure — Method.

vaporpressure(T)

Compute vapour pressure of water at the temperature T (degree Celsius) in hPa using Tetens equations. The temperature must be postive.

Monteith, J.L., and Unsworth, M.H. 2008. Principles of Environmental Physics. Third Ed. AP, Amsterdam. http://store.elsevier.com/Principles-of-Environmental-Physics/John-Monteith/isbn-9780080924793

source

Filtering

PhysOcean.gausswin — Function.

gausswin(N, α = 2.5)

Return a Gaussian window with N points with a standard deviation of (N-1)/(2 α).

source

PhysOcean.gaussfilter — Method.

gaussfilter(x,N)

Filter the vector x with a N-point Gaussian window.

source

Data download

PhysOcean.CMEMS.download — Function.

CMEMS.download(lonr,latr,timerange,param,username,password,basedir[; indexURLs = ...])

Download in situ data within the longitude range lonr (an array or tuple with two elements: the minimum longitude and the maximum longitude), the latitude range latr (likewise), time range timerange (an array or tuple with two DateTime structures: the starting date and the end date) from the CMEMS (Copernicus Marine environment monitoring service) in situ service [1]. param is one of the parameter codes as defined in [2] or [3]. username and password are the credentials to access data [1] and basedir is the directory under which the data is saved. indexURLs is a list of the URL to the index_history.txt file. Per default, it includes the URLs of the Baltic, Arctic, North West Shelf, Iberian, Mediteranean and Black Sea Thematic Assembly Center.

As these URLs might change, the latest version of the URLs to the indexes can be obtained at [1].

Example

julia> username = "..."
julia> password = "..."
julia> lonr = [7.6, 12.2]
julia> latr = [42, 44.5]
julia> timerange = [DateTime(2016,5,1),DateTime(2016,8,1)]
julia> param = "TEMP"
julia> basedir = "/tmp"
julia> files = CMEMS.download(lonr,latr,timerange,param,username,password,basedir)

[1]
http://marine.copernicus.eu/

[2]
http://www.coriolis.eu.org/Documentation/General-Informations-on-Data/Codes-Tables

[3]
http://doi.org/10.13155/40846

source

PhysOcean.CMEMS.load — Function.

data,lon,lat,z,time,ids = load(T,fname::TS,param; qualityflags = [good_data, probably_good_data]) where TS <: AbstractString

source

data,lon,lat,z,time,ids = CMEMS.load(T,fnames,param; qualityflags = ...)

Load all data in the vector of file names fnames corresponding to the parameter param as the data type T. Only the data with the quality flags CMEMS.good_data and CMEMS.probably_good_data are loaded per default. The output parameters correspondata to the data, longitude, latitude, depth, time (as DateTime) and an identifier (as String).

See also CMEMS.download.

source